LayoutTests/printing/printing-events-expected.txt

Basic testing for beforeprint / afterprint events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


window.print()
* Received beforeprint event
* Received afterprint event
PASS receivedBeforePrintEvent is true
PASS successfullyParsed is true

TEST COMPLETE
